AWS: Allow to specify an install group.

+++ src/aws/afterinstall_server.sh
@@ -12,6 +12,7 @@ fi
 [ -z $emobility_landscape ] && { echo "emobility landscape env variable not found, exiting"; exit 1; }
 [ -z $emobility_service_type ] && { echo "emobility env service type variable not found, exiting"; exit 1; }
 [ -z $emobility_user ] && { echo "emobility user env variable not found, exiting"; exit 1; }
+[ -z $emobility_group ] && { echo "emobility group env variable not found, exiting"; exit 1; }
 
 [ -e /etc/nginx/sites-enabled/default ] && sudo rm /etc/nginx/sites-enabled/default
 
@@ -19,4 +20,4 @@ sudo cp $emobility_install_dir/dist/assets/configs-aws/nginx-$emobility_service_
 sudo ln -sf /etc/nginx/sites-available/default_emobility /etc/nginx/sites-enabled/
 sudo systemctl reload nginx
 cp $emobility_install_dir/dist/assets/configs-aws/config-$emobility_service_type-$emobility_landscape.json $emobility_install_dir/dist/assets/config.json
-sudo chown $emobility_user.$emobility_user $emobility_install_dir
+sudo chown $emobility_user.$emobility_group $emobility_install_dir
